refactor: move budgeting::budget_over_tree types from lib to wire
