+++ lib/src/budgeting/budget_over_tree.rs
@@ -2,47 +2,10 @@ use crate::common::{request, SerializableNone};
 use crate::error::ApiError;
 use anyhow::Context;
 use chrono::{DateTime, FixedOffset};
+use lrzcc_wire::budgeting::BudgetOverTree;
 use reqwest::blocking::Client;
 use reqwest::{Method, StatusCode, Url};
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 use std::rc::Rc;
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct BudgetOverTreeServer {
-    pub total: f64,
-    pub flavors: HashMap<String, f64>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct BudgetOverTreeUser {
-    pub cost: f64,
-    pub budget_id: u32,
-    pub budget: u64,
-    pub over: bool,
-    pub servers: HashMap<String, BudgetOverTreeServer>,
-    pub flavors: HashMap<String, f64>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct BudgetOverTreeProject {
-    pub cost: f64,
-    pub budget_id: u32,
-    pub budget: u64,
-    pub over: bool,
-    pub users: HashMap<String, BudgetOverTreeUser>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub flavors: Option<HashMap<String, f64>>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct BudgetOverTree {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cost: Option<f64>,
-    pub projects: HashMap<String, BudgetOverTreeProject>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub flavors: Option<HashMap<String, f64>>,
-}
 
 pub struct BudgetOverTreeApi {
     pub url: String,
